requirements.txt
fabric
pySecurityCenter
SQLAlchemy